vroom-package {vroom}	R Documentation

vroom: Read and Write Rectangular Text Data Quickly

Description

The goal of 'vroom' is to read and write data (like 'csv', 'tsv' and 'fwf') quickly. When reading it uses a quick initial indexing step, then reads the values lazily , so only the data you actually use needs to be read. The writer formats the data in parallel and writes to disk asynchronously from formatting.

Author(s)

Maintainer: Jennifer Bryan jenny@rstudio.com (ORCID)

Authors:

Jim Hester jim.hester@rstudio.com (ORCID)

Hadley Wickham hadley@rstudio.com (ORCID)

Other contributors:

https://github.com/mandreyel/ (mio library) [copyright holder]

Jukka Jylänki (grisu3 implementation) [copyright holder]

Mikkel Jørgensen (grisu3 implementation) [copyright holder]

RStudio [copyright holder, funder]

See Also

Useful links:

https://vroom.r-lib.org

https://github.com/r-lib/vroom

Report bugs at https://github.com/r-lib/vroom/issues

[Package vroom version 1.5.7 Index]