vroom_altrep_opts {vroom}	R Documentation

Show which column types are using Altrep

Description

This function is deprecated in favor of vroom_altrep().

Usage

vroom_altrep_opts(which = NULL)

Arguments

which
A character vector of column types to use Altrep for. Can also take TRUE or FALSE to use Altrep for all possible or none of the types

[Package vroom version 1.6.4 Index]