Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>            X1 X2       X3            X4     X5
#>         <int> <time>   <time>     <dbl>  <dbl>
#>  1  318823166 06:16:13 18:58:37 -0.0209  0.931
#>  2 1905125835 11:34:15 03:02:56 -1.74    1.46 
#>  3  612526761 17:04:47 12:19:29  0.609   1.04 
#>  4 1221882361 21:22:54 05:33:28 -0.760   0.974
#>  5  185255979 12:42:48 16:48:37 -0.577  -0.312
#>  6 1739342133 22:21:22 17:52:47  0.725   0.209
#>  7  717553684 06:48:23 04:37:05 -1.20    0.483
#>  8  575355069 00:28:39 07:04:46  0.192  -0.279
#>  9 1301648262 08:57:09 01:06:58 -0.436   0.275
#> 10 1739476070 20:24:19 18:34:17 -2.02   -3.09 

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2      X3      X4
#>      <dbl>   <dbl>   <dbl>   <dbl>
#>  1 -1.85   -0.626  -0.161  -1.04  
#>  2 -0.962  -0.0195  0.666  -2.81  
#>  3 -0.0721 -1.20   -0.0461 -1.15  
#>  4  0.574  -1.16    1.57    0.762 
#>  5 -0.430  -0.547   0.342  -0.230 
#>  6  0.602  -0.376   0.181   0.499 
#>  7 -1.18   -0.534  -1.62    0.977 
#>  8  1.94   -0.574   1.41   -1.42  
#>  9  0.520  -1.08    1.01   -0.0818
#> 10  0.483  -1.14    0.162  -0.505 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1    X2    X3    X4
#>     <dbl> <dbl> <dbl> <dbl>
#>  1  3.44  -3.65  8.18 15.4 
#>  2  7.50  11.3   8.15 16.4 
#>  3 -6.00  24.7  15.8  11.9 
#>  4  0.168  1.04 -9.17 -1.84
#>  5 -9.13  -1.81 -7.65 -6.22
#>  6  7.26  24.9   6.98 -6.78
#>  7 -7.66  -3.27 -8.66 14.9 
#>  8 22.4    5.57  7.76 23.5 
#>  9  0.955 -8.81 -8.10 10.5 
#> 10 17.6   24.6  20.1  -6.04
#> # … with 15 more rows