Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                      X2         X3      X4 X5             
#>    <dttm>               <dbl>      <int>   <dbl> <fct>          
#>  1 2002-06-12 21:15:51  1.17   534980635 -0.236  last_hartebeest
#>  2 2012-02-05 20:31:15  0.348  195404736  0.723  warm_kitten    
#>  3 2017-10-05 05:01:24  0.309   11572591  0.522  last_hartebeest
#>  4 2018-10-26 22:21:27 -0.395  712850907 -0.0308 last_hartebeest
#>  5 2004-02-19 18:42:30  0.455  866617048 -1.60   wet_weasel     
#>  6 2019-01-10 21:56:56 -0.302 1013098869  1.08   warm_kitten    
#>  7 2008-11-03 23:53:32  1.35  1090379568  1.98   wet_weasel     
#>  8 2020-07-23 08:49:32  0.167 1397163839 -1.00   warm_kitten    
#>  9 2008-05-09 17:10:50  0.753 2055817630  2.88   warm_kitten    
#> 10 2017-01-03 01:36:54  0.621   87572993  0.972  wet_weasel     

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1      X2      X3     X4
#>     <dbl>   <dbl>   <dbl>  <dbl>
#>  1  1.39   1.66    0.573   0.191
#>  2  0.373  2.68    0.620   0.811
#>  3 -0.727 -1.48   -0.138   1.16 
#>  4  0.308 -0.215  -1.83    1.52 
#>  5 -0.713  0.288   1.38   -0.796
#>  6  0.844 -0.881  -0.0373 -0.596
#>  7 -0.612 -0.710   1.15   -0.665
#>  8  1.27  -0.732   0.616   0.675
#>  9  1.34   0.0615  0.620   0.353
#> 10 -0.612 -1.03    0.436  -0.581
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2     X3     X4
#>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -1.34 10.2   14.2   -7.88 
#>  2 18.6  16.5    5.69  21.6  
#>  3 -8.83  3.72   4.78  24.1  
#>  4 22.4   9.34  -0.390  4.74 
#>  5 -2.06 10.4    1.34  17.8  
#>  6  3.60 -4.64  -1.24  20.9  
#>  7 -3.50  8.37  -6.34  -0.752
#>  8 10.8  20.8   15.8   17.3  
#>  9  2.24 -0.471 -2.04  16.2  
#> 10 11.3  -2.56  24.4    3.34 
#> # … with 15 more rows