Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>        X1         X2 X3                        X4              X5         
#>     <dbl>      <int> <chr>                     <fct>           <chr>      
#>  1  0.554  313069684 S8Pb8                     ugliest_gazelle Fa0WN2vgu5…
#>  2  0.211  834515334 pLndQ2b6mQVb              beautiful_hyena AaKg6JP1Yz…
#>  3  0.878  494751137 zmAx06                    broken_tapir    jHonLTFknP…
#>  4 -0.653 2062762285 vCFSZ24KQmtMJv7nyHnIftkYn broken_tapir    nbCv2ZUrNvF
#>  5  0.132  415093330 cM1Y9975zyc               broken_tapir    uS11JH2EHs…
#>  6  0.288 1033566413 7uCfndtwdfv               delightful_bee  8QRwFQ9    
#>  7  1.04   201969288 cVuQ6jBHGx3RzxRubGqjA8u   full_mouse      eQ4FUikSNt…
#>  8  1.11   566845823 fyhMx4dyoQOeqdqNevbLc     icy_weasel      Hqckv      
#>  9  1.41  1964110700 iLXhcJmL                  beautiful_hyena gZuNzZSjBG…
#> 10 -1.08  2064597622 ZakfkekYbGXUBljceeqRBH1   lazy_canary     2c2Tvx2IwR 

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1     X2     X3      X4
#>     <dbl>  <dbl>  <dbl>   <dbl>
#>  1  2.05   0.437 -0.852  0.311 
#>  2  1.72  -1.54  -0.970  0.0946
#>  3 -0.532  1.02   0.454 -0.128 
#>  4  0.697 -0.628  0.916  1.43  
#>  5  1.57   0.741  0.593  0.592 
#>  6 -1.79   0.586  0.165  0.115 
#>  7 -0.440  0.808  0.773 -0.573 
#>  8 -1.24   2.21   0.530 -0.728 
#>  9 -0.665  1.87   0.274 -0.456 
#> 10  0.334  0.445  1.04  -0.725 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1    X2    X3    X4
#>     <dbl> <dbl> <dbl> <dbl>
#>  1  4.43  -5.94 11.1  23.9 
#>  2 15.7    9.80  5.90 -3.66
#>  3 23.4    7.09  2.19  2.84
#>  4 -0.855  1.85 -8.05 18.2 
#>  5 10.2   15.1  16.6   5.03
#>  6 15.5   -9.54 18.9   6.17
#>  7 12.2    6.74 -3.15 22.9 
#>  8 22.9    4.67 16.0  -1.74
#>  9  3.17   4.33 -8.34 24.5 
#> 10 16.7   20.3   4.71 -1.67
#> # … with 15 more rows