Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1         X2                X3                 X4     X5
#>    <date>     <fct>             <date>          <int>  <dbl>
#>  1 2006-08-15 gigantic_giraffe  2018-01-27  457797850  1.46 
#>  2 2007-09-14 tender_coyote     2016-06-24  717755064 -0.548
#>  3 2017-05-08 chilly_marten     2012-05-10  258603093  1.18 
#>  4 2015-05-31 powerful_lovebird 2015-04-11  119426524  0.889
#>  5 2018-05-29 gigantic_giraffe  2019-01-12 1637696062 -0.638
#>  6 2014-04-20 chilly_marten     2015-05-22 1828448641 -0.229
#>  7 2012-10-15 loud_meerkat      2009-03-10 1412126456  0.883
#>  8 2003-08-02 inexpensive_tiger 2003-10-03    6415858 -1.33 
#>  9 2002-09-20 red_addax         2006-01-17 1097085974  0.314
#> 10 2020-08-04 uneven_budgerigar 2011-11-08 1681382618  1.05 

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1       X2      X3      X4
#>      <dbl>    <dbl>   <dbl>   <dbl>
#>  1  0.880  -0.596    0.622   0.806 
#>  2  0.784   0.686    0.528   0.804 
#>  3  0.537  -1.30     1.24   -0.547 
#>  4 -0.929   0.0743  -0.342  -0.949 
#>  5 -0.0716  0.532    0.818   0.696 
#>  6  0.528   0.200    0.690   0.501 
#>  7 -1.01    0.276    0.435   0.269 
#>  8  1.26    0.542    0.0369 -0.280 
#>  9  0.0278 -0.00934  0.828   0.831 
#> 10  0.375  -0.773    0.0172  0.0265
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1    X2    X3      X4
#>     <dbl> <dbl> <dbl>   <dbl>
#>  1 -6.66  17.8  -3.05  6.24  
#>  2 -1.67  11.7  -2.56  8.66  
#>  3 21.0   14.9  -8.79 10.2   
#>  4 24.6   15.9   3.25  2.55  
#>  5 -0.488  6.04 19.7   5.43  
#>  6 -2.85   7.97 -1.48  2.17  
#>  7 24.5   12.9  15.6  21.5   
#>  8  0.464 17.0  16.0  11.6   
#>  9  5.26   8.23 18.2  -0.0434
#> 10 15.6    4.81 22.8  -2.89  
#> # … with 15 more rows