Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>            X1      X2      X3         X4 X5        
#>         <int>   <dbl>   <dbl>      <int> <date>    
#>  1  159264844 -1.53   -0.0758 1337261232 2007-02-13
#>  2 1907048974  0.983  -0.879   716333606 2008-05-15
#>  3 1014761514 -0.0618 -0.0736 1525573889 2009-10-26
#>  4 1564460678 -0.469  -2.49   2027259002 2002-05-10
#>  5 1040252406  1.02    0.0424 1087493159 2010-11-22
#>  6  880776381  1.17   -0.0355 1401089978 2011-04-16
#>  7 1518718363  0.0765 -0.174   728936601 2018-01-04
#>  8 1912974951  1.38    0.168    44978739 2013-08-19
#>  9  501164689 -0.737   0.124  1542997388 2005-08-28
#> 10 1929080290  0.366   1.20    568469726 2007-05-27

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>          X1     X2      X3     X4
#>       <dbl>  <dbl>   <dbl>  <dbl>
#>  1  0.709   -0.897 -0.836   2.17 
#>  2  0.191    0.344 -1.52    0.465
#>  3  0.684   -2.03   0.978  -0.707
#>  4 -0.304   -0.179  1.80   -0.279
#>  5  0.969   -1.45  -0.0472 -1.29 
#>  6  0.508    0.685 -0.391  -0.251
#>  7 -0.00656 -0.755  0.334  -0.458
#>  8 -0.254    0.964  0.623  -0.744
#>  9 -0.517   -0.281 -0.719   1.00 
#> 10  1.74     1.38   1.75    1.54 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1    X2    X3    X4
#>     <dbl> <dbl> <dbl> <dbl>
#>  1 17.3   11.3  -2.78 14.8 
#>  2 18.9   14.5  10.1  -9.13
#>  3 -3.41  12.8  10.2  -1.99
#>  4  1.30  -9.85 13.1  12.3 
#>  5 17.2    8.58 -1.53 14.4 
#>  6 21.6    9.89  5.33 19.0 
#>  7 -5.54  13.1  19.3   8.56
#>  8 21.3   19.2   7.53  9.74
#>  9 11.1   20.7  -3.80  3.25
#> 10  0.135  6.97 14.3  12.8 
#> # … with 15 more rows