Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>        X1 X2                  X3                   X4         X5
#>     <dbl> <dttm>              <fct>             <dbl>      <int>
#>  1 -1.27  2013-10-22 16:33:28 last_beaver       0.128 1414987380
#>  2  0.166 2013-02-08 10:22:46 black_walrus      1.17   211552729
#>  3  0.539 2014-05-10 21:30:04 ancient_hedgehog  0.134  498625833
#>  4  0.738 2015-12-06 05:23:56 black_walrus     -1.06  1357108000
#>  5  0.773 2007-12-04 08:56:53 black_walrus      1.29   109272593
#>  6  1.10  2011-03-23 20:22:56 obnoxious_eagle  -0.941 1361497174
#>  7 -0.672 2001-10-15 20:37:35 greasy_basilisk  -0.810  560031106
#>  8  0.541 2015-11-27 00:05:24 small_crab        0.533  615665905
#>  9  1.16  2009-03-18 13:40:00 helpful_lovebird  0.458  711005752
#> 10  0.924 2002-12-13 19:24:54 boiling_hog       2.03  1828130699

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2      X3      X4
#>      <dbl>   <dbl>   <dbl>   <dbl>
#>  1 -0.284   0.0574 -2.12    0.361 
#>  2  0.0410 -2.33   -0.0260  1.94  
#>  3 -0.0275 -0.0221  1.16    0.520 
#>  4 -0.589   0.903   0.408  -0.504 
#>  5  0.854   0.549  -0.0141  0.185 
#>  6  0.542   0.248   1.13   -0.777 
#>  7  0.757   0.232  -0.655   0.175 
#>  8  0.159  -1.13   -0.898  -0.150 
#>  9  0.426   0.179   0.734  -0.0316
#> 10  0.483   0.673  -0.578   0.967 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1      X2    X3     X4
#>    <dbl>   <dbl> <dbl>  <dbl>
#>  1 -9.94  2.58   15.5  16.0  
#>  2 11.9  19.1    22.1  -5.19 
#>  3  2.72 24.6     3.71 10.4  
#>  4 -9.76 11.6     9.30 17.6  
#>  5 -7.32 -0.0619 20.4  10.8  
#>  6 11.2   9.04    1.04  3.64 
#>  7 18.1  23.1     7.15 -0.508
#>  8 20.6   2.25   -3.90  1.19 
#>  9 -2.12  5.74    1.14 21.9  
#> 10  6.58 24.5    23.8   0.231
#> # … with 15 more rows