Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage,

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                  X2         X3       X4         X5      
#>    <dttm>              <date>     <time>   <date>     <time>  
#>  1 2020-01-06 08:25:54 2011-12-11 06:15:26 2020-08-30 16:52:48
#>  2 2004-10-29 18:41:31 2007-01-08 04:32:58 2016-05-23 10:24:05
#>  3 2019-10-23 09:48:38 2005-01-29 02:30:51 2014-08-21 03:21:53
#>  4 2002-08-31 02:28:01 2011-02-18 11:42:17 2016-10-26 21:09:38
#>  5 2012-05-19 05:15:52 2014-10-11 22:01:00 2017-07-15 18:07:01
#>  6 2004-03-17 23:33:32 2012-06-15 06:33:27 2018-09-25 03:09:10
#>  7 2006-03-06 14:04:39 2017-09-30 05:36:59 2019-10-06 04:29:53
#>  8 2013-04-19 16:27:38 2016-07-19 20:48:00 2019-04-20 01:23:55
#>  9 2006-11-11 15:07:32 2003-05-10 06:00:42 2016-03-16 11:59:59
#> 10 2002-08-05 12:47:27 2008-08-11 18:46:40 2013-11-29 02:30:01

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1      X2      X3     X4
#>     <dbl>   <dbl>   <dbl>  <dbl>
#>  1  0.469 -1.16   -0.374   0.102
#>  2  0.531 -0.456  -2.20   -2.04 
#>  3  0.657  0.575  -0.579  -1.10 
#>  4 -0.130 -0.482   1.99    0.584
#>  5  0.681  1.25    1.33    0.269
#>  6 -0.814  0.598  -0.0266 -0.174
#>  7 -1.61  -0.893   0.583   0.333
#>  8 -1.11   0.371   1.16    0.225
#>  9 -0.972  0.0334 -0.675   0.301
#> 10  0.206 -0.393   1.23    1.16 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2     X3    X4
#>    <dbl>  <dbl>  <dbl> <dbl>
#>  1 11.1  17.4    9.71   2.38
#>  2  7.75 15.4   17.4    9.23
#>  3 -1.68 17.1    6.41   8.21
#>  4 -9.61  5.40  19.8   23.1 
#>  5 16.5  24.6    5.19  -6.36
#>  6 -2.07 -0.799 10.4   -7.44
#>  7  4.94 13.2   19.6   19.0 
#>  8  4.72  5.33  18.8   14.2 
#>  9  4.22 10.4   -0.257 -1.40
#> 10  2.41 -9.71  13.1    3.41
#> # … with 15 more rows