Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                         X2 X3                    X4 X5      
#>    <chr>                   <dbl> <fct>              <int> <time>  
#>  1 wMHjQzx7LSqI5g3cKIP2    0.653 raspy_ferret  1502257856 04:56:59
#>  2 bESRS3iBkdLOcBHJvfL2P   0.222 uptight_goat   996668138 21:32:01
#>  3 11Z2IwL9tKg0soVs8eW    -0.296 quick_ocelot   931410082 07:05:20
#>  4 XaHUmYUCU9LCHrJlez1Hsq  2.32  bad_beaver    1217977107 12:05:00
#>  5 hHlJeSwSwWDxWnblPj      0.629 short_musk-ox 1726964037 17:36:11
#>  6 fr3DPL3sBpzvbKyUnhyxi  -0.388 raspy_ferret  1479037674 23:27:56
#>  7 l45IywV2IS5LfY          0.495 red_frog      1605998693 07:43:24
#>  8 XmsAiuYm3DzPD0iP7SYCh1  0.701 short_musk-ox 1654846695 03:06:14
#>  9 1N895                  -0.222 quick_ocelot   782806220 23:01:13
#> 10 fmNMhNUPU5H             0.555 raspy_ferret   346814458 19:11:45

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1     X2     X3     X4
#>     <dbl>  <dbl>  <dbl>  <dbl>
#>  1  1.16  -0.585  0.262 -1.13 
#>  2 -1.14  -0.158 -0.706 -0.122
#>  3 -0.949  1.52   0.601  0.672
#>  4 -0.552 -0.174  0.492 -0.841
#>  5 -0.253  1.81  -0.112 -1.03 
#>  6 -1.34  -1.16   1.03  -1.78 
#>  7  1.01   1.59   1.29   1.77 
#>  8 -0.884 -1.30   1.26  -1.21 
#>  9  1.14  -0.303 -0.374  1.57 
#> 10  0.758 -0.193  0.365 -1.25 
#> # … with 15 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2    X3    X4
#>    <dbl>  <dbl> <dbl> <dbl>
#>  1  8.15  2.39  14.6  -7.03
#>  2 23.6  16.3   14.1  -5.80
#>  3  2.17 -0.566 19.6   5.80
#>  4 24.3   4.54  15.5  23.6 
#>  5 -9.28 16.4   24.2  21.4 
#>  6 23.6  18.7   19.3  21.9 
#>  7  1.60 -2.59  19.8   4.87
#>  8 24.0   6.00  -7.57 15.6 
#>  9 -1.41 -9.65  -1.08 16.3 
#> 10 -1.78 19.8   15.5   7.67
#> # … with 15 more rows
#> # ℹ Use `print(n = ...)` to see more rows