Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                  X2                      X3 X4                    X5
#>    <dttm>              <fct>                <int> <fct>              <int>
#>  1 2012-10-19 02:21:01 short_lynx      1861577128 strong_impala     1.48e9
#>  2 2008-08-28 12:45:57 sticky_alpaca    311907661 many_eland        1.04e9
#>  3 2009-02-13 11:02:05 shallow_sheep    379920991 thankful_deer     1.52e9
#>  4 2012-09-20 02:38:13 puny_mule        528413641 jealous_coati     1.78e9
#>  5 2002-04-15 13:26:02 better_tapir    1248461564 silly_dromedary   6.31e8
#>  6 2020-08-08 14:17:57 noisy_dog        130947778 wooden_mare       1.31e9
#>  7 2005-10-21 17:54:00 fluffy_impala    310188586 quaint_seal       8.19e8
#>  8 2004-10-13 07:41:07 panicky_buffalo  927258376 scary_crow        1.61e9
#>  9 2017-11-22 22:44:25 damaged_finch   1209672493 cool_jaguar       8.73e8
#> 10 2008-01-24 21:12:44 fluffy_impala     26895443 cool_jaguar       7.04e8

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1     X2     X3     X4
#>      <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -0.811  -0.649  2.00   1.28 
#>  2  0.0614  0.167  0.498  1.31 
#>  3 -0.345  -0.550 -0.114  0.702
#>  4 -0.709   1.69   0.360  0.135
#>  5  0.0257  0.298 -1.12  -2.05 
#>  6  0.816   1.35  -0.183 -0.565
#>  7  0.761   0.786 -0.354  0.432
#>  8  0.641   0.691  0.108 -1.73 
#>  9  0.692   0.567  0.726  0.438
#> 10 -0.0106 -0.834 -0.960  1.06 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2    X3    X4
#>    <dbl>  <dbl> <dbl> <dbl>
#>  1 -9.59 15.8    4.49 -7.47
#>  2 22.1   0.691 16.7  -9.93
#>  3 15.5   4.12   5.89  3.78
#>  4  6.35 23.6   -5.56 -6.51
#>  5 16.6   0.910 11.9  -1.70
#>  6 -8.19 10.1   -8.30 -7.99
#>  7 -3.67 17.2   -9.99 11.9 
#>  8 20.3  -2.48  -8.04  3.77
#>  9 22.7  -4.58  -2.43 -8.10
#> 10 -7.83 20.8   -2.03 13.6 
#> # … with 15 more rows