Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>            X1         X2 X3                                X4 X5          
#>         <int>      <int> <chr>                          <int> <fct>       
#>  1 1100914402  214297585 c9ZsU6AxxwTfvMnz2gJ        903336986 gentle_snake
#>  2 1813115031 1167429229 LSZVLOtd5ggkdij47UhU      1888314569 gentle_snake
#>  3 1288869786   14725380 JaoKnOiW                  1930800982 scary_lynx  
#>  4  163776987 1746621894 euu6SZ5NacS9r3dOenNjLf2   1816941764 gentle_snake
#>  5  836177806 1942752456 v2JAHmKejV6b32PzjLvfvpe     99016498 scary_lynx  
#>  6  574517486  154694850 MYgcSD6yd1SWOEh           2007576985 gentle_snake
#>  7  714121456  749202831 f5kzX0bCc70OWF            1016900995 scary_lynx  
#>  8 1290023163 1031171347 E9RfBdMC5KjYsZeW4zU       1242991200 scary_lynx  
#>  9   99840194  985100169 iQdrqajs3                  761444703 gentle_snake
#> 10  113450528 2029247513 7V5KmeJYQ20MOuZfzHszKyDxk 1544719172 scary_lynx  

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1     X2     X3      X4
#>     <dbl>  <dbl>  <dbl>   <dbl>
#>  1  0.878  1.44  -0.433 -0.510 
#>  2  0.209 -1.35   0.963  1.15  
#>  3 -0.652 -0.775  1.24   1.03  
#>  4  1.10   0.693 -0.144 -0.974 
#>  5 -0.869 -1.70  -1.49   1.06  
#>  6  0.926  0.521  0.192  0.222 
#>  7  1.00   0.312  0.626  0.0718
#>  8 -1.03   0.563  0.790  2.06  
#>  9  0.347  1.27  -1.67  -0.918 
#> 10 -0.284  0.396  1.25  -0.434 
#> # ℹ 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1    X2     X3    X4
#>    <dbl> <dbl>  <dbl> <dbl>
#>  1 -5.08  4.76 22.6   19.6 
#>  2  3.26 19.0  12.6   -7.03
#>  3 17.1  18.9  24.2   21.6 
#>  4 17.0   5.57 -9.04  20.2 
#>  5 -2.33 18.6  16.3   -4.91
#>  6 12.1  16.1   6.37  24.8 
#>  7  4.69 11.5   1.71  -9.78
#>  8 -1.38 -1.79 -0.533 -7.02
#>  9 11.1  -8.27 17.3   -2.16
#> 10  3.62 23.9  -6.54  17.1 
#> # ℹ 15 more rows