Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                X2                          X3 X4           X5       
#>    <fct>             <dttm>                   <int> <fct>        <fct>    
#>  1 warm_elk          2009-09-16 19:04:48  782057532 white_vicuna helpful_…
#>  2 cooing_iguana     2010-07-19 01:37:08 2136536638 dry_opossum  quiet_sh…
#>  3 obedient_tiger    2014-06-13 16:18:33  653696187 white_vicuna heavy_ch…
#>  4 clever_basilisk   2009-04-23 04:57:19 2050039892 white_vicuna odd_sloth
#>  5 tasteless_koala   2002-07-13 14:19:56  780963771 dry_opossum  quiet_sh…
#>  6 tasteless_koala   2013-01-06 10:54:00 1566070744 white_vicuna odd_sloth
#>  7 miniature_chicken 2003-06-09 00:51:23  956676266 white_vicuna heavy_ch…
#>  8 warm_elk          2014-09-04 21:02:39  857230026 white_vicuna hot_cana…
#>  9 clever_basilisk   2018-07-09 23:25:02 1740702558 white_vicuna helpful_…
#> 10 eager_badger      2001-06-26 16:33:25  104160091 dry_opossum  heavy_ch…

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1       X2      X3     X4
#>      <dbl>    <dbl>   <dbl>  <dbl>
#>  1 -1.66   -0.535    1.79    0.239
#>  2 -0.712  -0.440    0.0608 -0.240
#>  3 -0.0954  1.81     0.573   0.498
#>  4 -0.520   2.20     1.05   -2.06 
#>  5 -0.770  -1.48     0.794  -0.151
#>  6  0.960   2.31     1.37    1.46 
#>  7  0.774  -2.70     1.89   -1.10 
#>  8 -0.996  -0.00915 -0.211  -2.19 
#>  9 -0.345  -0.00290  0.282  -1.81 
#> 10 -0.986   0.315    0.417  -0.570
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2     X3    X4
#>    <dbl>  <dbl>  <dbl> <dbl>
#>  1 -2.82  4.23   5.41   7.61
#>  2 20.3  -0.190 -5.93  -9.01
#>  3  7.59 16.6   21.2   18.1 
#>  4 23.0  23.2   16.0    2.55
#>  5 19.7  -8.09   2.83  24.6 
#>  6 -2.55 18.2   -3.23  -3.44
#>  7 18.2  14.2    0.178 13.0 
#>  8 13.8  -6.09  10.9   -6.10
#>  9 11.5   0.335 -4.19  -5.73
#> 10 21.5  15.1    0.939 10.7 
#> # … with 15 more rows