Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>        X1 X2                    X3         X4                  X5
#>     <dbl> <chr>                 <date>     <fct>            <dbl>
#>  1 -0.144 AHsihj                2012-01-14 handsome_gopher  0.329
#>  2 -2.38  oCZLfkD8Bq7CYpsaPuFD7 2003-01-20 curly_toad       0.457
#>  3 -1.70  bMHNCn                2002-11-16 handsome_mole   -0.337
#>  4 -0.619 ECziL436C3pJyIV9U     2001-05-19 handsome_mole   -1.20 
#>  5  0.508 tbCpd6tcpoATFka       2001-07-03 odd_llama       -1.44 
#>  6 -1.26  bPTCZuNTM6itmntLCteKf 2015-09-25 weak_meerkat    -0.664
#>  7 -0.921 baTsqr7Ye0ptde0Z      2009-11-26 handsome_gopher -0.428
#>  8 -1.43  7sEQGaKjO5T1VkhAasAgt 2012-08-09 handsome_mole   -0.625
#>  9  1.01  gjc4RCb7COJTJZf       2001-02-06 drab_koala      -1.09 
#> 10  0.497 G7BNqilG3m1OY         2016-10-28 yummy_peccary   -0.810

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2      X3     X4
#>      <dbl>   <dbl>   <dbl>  <dbl>
#>  1  0.854   1.79    0.482   0.567
#>  2  1.32   -1.56   -1.04   -0.235
#>  3 -0.943  -0.414  -0.0933  0.189
#>  4  0.906  -0.536  -0.362  -0.446
#>  5 -1.28    0.545  -0.0729  0.758
#>  6 -0.468  -1.52    0.550  -0.182
#>  7 -0.0373  1.84   -0.775  -0.281
#>  8  0.841   1.21    0.0210 -0.306
#>  9 -0.484  -0.0332  0.284   1.44 
#> 10 -1.41   -1.80    0.368   2.41 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1      X2     X3     X4
#>     <dbl>   <dbl>  <dbl>  <dbl>
#>  1  4.34  11.2     0.769 -9.49 
#>  2 17.5    0.0853 18.9   -7.14 
#>  3 18.7    2.73    8.71   7.99 
#>  4 23.5   13.3     7.40   6.79 
#>  5 18.8    6.73   -7.69  -2.26 
#>  6  9.62  19.1    18.7   -3.69 
#>  7 14.3   -3.35   -5.76  16.6  
#>  8 -0.898 22.8    -1.16  -4.25 
#>  9  6.55   5.53   16.2   -0.372
#> 10  9.17   6.80   18.2   18.9  
#> # … with 15 more rows