Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1       X2                       X3            X4 X5                 
#>    <time>   <chr>                    <time>     <dbl> <dttm>             
#>  1 07:38:37 IJfMllDiW8NrLcuNa        20:06:16 -0.292  2012-10-10 23:19:19
#>  2 08:22:52 YSDjZlavJ1jZHg9svsId7BtL 22:13:31 -1.37   2017-01-05 09:26:12
#>  3 14:39:40 taOyJYKPOHTZTLpXlMu3     07:49:58  0.201  2013-10-07 13:48:27
#>  4 03:12:31 CaoLZrLZdfOh4RYlPBKmt    08:22:17 -0.0326 2011-03-17 17:17:53
#>  5 04:10:23 sTAOBO3JxToiknlYljMTN9   03:02:21  0.686  2006-05-17 10:10:31
#>  6 07:34:46 nFtT62hiiiefCDC0         15:47:22  0.301  2011-03-11 20:00:28
#>  7 16:07:27 5r4ZEnjlyh1o             11:06:59  0.0971 2019-02-24 07:48:40
#>  8 09:30:13 z3VDNfzM79QBn3OA7        21:23:23  0.614  2003-04-18 17:19:51
#>  9 01:29:37 7u8EvEjtk0uuIC2jg8Zq     18:41:27 -1.76   2014-12-09 21:29:12
#> 10 02:25:07 AkffP6TIorQAS            23:15:05  0.209  2019-11-20 17:40:52

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1     X2      X3      X4
#>      <dbl>  <dbl>   <dbl>   <dbl>
#>  1  1.55    0.279  1.71   -0.107 
#>  2 -0.190   0.741  0.793  -0.0197
#>  3  1.26    1.84   1.16    0.366 
#>  4  0.198  -0.419 -1.28   -0.0453
#>  5  0.205  -1.23   1.99   -0.534 
#>  6 -0.700  -1.01   0.745   0.775 
#>  7 -0.894  -0.237  0.0961  2.10  
#>  8  0.291  -1.43   0.219  -0.0398
#>  9  0.0302  0.614  2.50   -0.718 
#> 10  0.617   0.156 -0.446   1.18  
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1      X2      X3    X4
#>     <dbl>   <dbl>   <dbl> <dbl>
#>  1  1.45   9.81    0.0987 -1.19
#>  2 23.4   10.0     7.93   24.6 
#>  3 -8.13  18.7    -5.63    1.53
#>  4  3.62  -1.93   18.5    21.0 
#>  5 -5.02   0.0915  5.53   15.8 
#>  6  0.372 10.2     2.57   -1.86
#>  7 -7.69  -1.78   -6.59   10.5 
#>  8  8.57  21.2    -1.80   -8.71
#>  9 -2.82   4.08   22.1    21.3 
#> 10 -8.63  18.9    15.1    17.6 
#> # … with 15 more rows