Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                            X2 X3       X4         X5               
#>    <chr>                      <dbl> <time>   <date>     <fct>            
#>  1 zC0bs5ByRQbwsNrIaI4lBkMKt -1.55  17:17:14 2013-02-11 silly_stallion   
#>  2 T4jzW8ckZTpCR              0.529 09:16:47 2018-08-31 wrong_porcupine  
#>  3 sIUQbiRq9RTIGUIzfW1sFq     1.10  01:07:47 2011-07-25 clever_ermine    
#>  4 jLG2kOSY37DZ9r52           0.304 04:39:08 2001-07-25 empty_dog        
#>  5 EqT4wClvE2KJRmE           -0.432 08:26:21 2005-05-22 gray_fish        
#>  6 VuHARUPYcbXrI60lkdiV       0.612 17:22:58 2016-08-08 obedient_elephant
#>  7 V0ZHDVYnmZ4Y7zROXZ         1.34  03:09:12 2014-09-14 empty_dog        
#>  8 lIAbECK1VinNNTb0Xw5X      -0.342 13:21:51 2005-05-05 clumsy_bear      
#>  9 apawPuziXVn3he            -0.818 02:08:59 2003-07-01 wrong_porcupine  
#> 10 L7kWZc7Rh                 -1.96  18:46:58 2009-08-21 straight_yak     

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2       X3     X4
#>      <dbl>   <dbl>    <dbl>  <dbl>
#>  1 -1.18    0.432  -0.362   -1.44 
#>  2 -0.774  -0.0302  0.620   -2.10 
#>  3 -0.450  -0.926   1.19     2.27 
#>  4 -0.398   0.695   0.00842 -2.30 
#>  5  0.0136 -3.06    0.837   -0.127
#>  6  0.818   0.771   0.839    0.777
#>  7  0.522  -0.300  -2.05    -1.21 
#>  8  0.622  -0.791  -1.70     2.44 
#>  9 -1.30    0.398   1.63     0.505
#> 10  0.465   0.282   0.764    0.237
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1    X2    X3     X4
#>     <dbl> <dbl> <dbl>  <dbl>
#>  1 11.3   21.7  19.1  14.6  
#>  2  6.97  23.3  19.4  18.0  
#>  3 13.8   21.5  -2.66 -7.81 
#>  4 21.4   -1.55  4.32 -2.40 
#>  5  7.62   8.06 25.0  24.5  
#>  6  0.779 17.0   4.79  2.72 
#>  7 -1.33  -8.93 16.9  -0.318
#>  8 11.1    7.73 -2.60 11.1  
#>  9 -9.77  23.1  12.3  24.4  
#> 10 -9.34  24.0  -8.15  6.18 
#> # … with 15 more rows