Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                  X2       X3       X4                 
#>    <dttm>              <time>   <time>   <dttm>             
#>  1 2010-10-30 13:15:03 12:39:35 07:13:33 2009-07-08 23:49:09
#>  2 2008-02-20 11:39:15 23:28:21 14:28:07 2003-09-28 21:13:46
#>  3 2010-11-30 18:45:51 20:34:15 06:12:00 2011-08-15 04:09:20
#>  4 2004-03-13 01:51:34 20:07:45 09:17:02 2014-12-20 10:55:00
#>  5 2011-09-21 14:02:23 02:02:41 21:31:18 2014-02-21 23:24:50
#>  6 2001-10-16 20:04:40 07:22:42 05:11:26 2016-01-22 20:45:42
#>  7 2005-04-21 18:43:15 23:45:37 21:44:43 2003-05-17 03:55:15
#>  8 2019-05-21 03:11:52 16:25:28 04:42:10 2003-05-10 21:09:12
#>  9 2013-08-06 22:19:05 03:43:23 20:39:48 2014-12-03 11:15:31
#> 10 2017-08-04 11:24:33 01:30:04 19:04:50 2005-02-21 18:26:28
#> # … with 1 more variable: X5 <dttm>

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1      X2      X3     X4
#>     <dbl>   <dbl>   <dbl>  <dbl>
#>  1  0.834  0.384  -0.511  -1.73 
#>  2  0.803 -0.404   1.45    0.644
#>  3 -0.125  0.362   0.477  -0.106
#>  4  0.750  0.0186  1.61    0.835
#>  5  0.307 -0.709   0.236  -0.575
#>  6  0.799 -0.918   0.239   0.367
#>  7 -1.26   0.765  -0.0652 -0.476
#>  8 -0.773 -0.133   0.387  -0.245
#>  9  1.13  -1.12   -0.374  -1.83 
#> 10 -0.635 -0.115  -0.0868  0.464
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1    X2    X3    X4
#>    <dbl> <dbl> <dbl> <dbl>
#>  1  8.46 24.2  -8.83  8.63
#>  2 -7.35 21.9  10.8  -4.00
#>  3 12.6  -5.18 -2.64  4.89
#>  4 21.9  15.4  17.6   8.01
#>  5 10.9   4.88 24.3   6.53
#>  6 16.1  10.5  24.6   7.83
#>  7  8.12 14.2  23.0  15.5 
#>  8 -4.80  9.69 -4.58 20.7 
#>  9 23.2  -3.64  8.85 -6.84
#> 10 22.6  -5.17 -5.91 10.7 
#> # … with 15 more rows