Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                          X2 X3                 X4      X5
#>    <dttm>                   <int> <date>          <int>   <dbl>
#>  1 2017-02-05 12:38:50  526236618 2012-12-08  777325711 -0.0898
#>  2 2015-01-24 22:29:57 1157795327 2013-03-15  731716758  0.209 
#>  3 2012-12-16 09:47:33  292317808 2019-10-27 1429411130  1.59  
#>  4 2018-07-02 06:29:04 1731649416 2017-10-09 1883901889 -0.145 
#>  5 2018-09-02 18:17:10 1925784149 2006-08-19 1474532991  0.275 
#>  6 2011-06-20 00:45:29  243963261 2020-12-30  524120784  0.421 
#>  7 2003-02-05 20:44:06  584043866 2018-06-01  568151673  0.322 
#>  8 2017-04-23 22:02:39  857390808 2002-10-03 2030341416  0.991 
#>  9 2004-03-09 22:59:50  788208195 2006-10-19  146246033 -0.230 
#> 10 2002-05-01 05:50:38 1208969607 2013-04-08 1959401343  0.216 

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2      X3     X4
#>      <dbl>   <dbl>   <dbl>  <dbl>
#>  1 -0.0508 -0.583  -0.101  -0.450
#>  2 -1.58    0.0812  0.513   0.999
#>  3 -2.47   -1.21   -1.61   -0.785
#>  4 -0.639   0.501  -0.0722  1.81 
#>  5 -1.76   -0.807   0.807  -1.46 
#>  6  3.11    0.368  -0.462   0.153
#>  7 -1.14   -0.367  -1.04   -1.18 
#>  8 -1.19    0.777  -0.987  -1.55 
#>  9 -0.407   0.580   0.229  -0.303
#> 10 -0.707   0.914  -0.255   1.27 
#> # ℹ 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1    X2    X3    X4
#>    <dbl> <dbl> <dbl> <dbl>
#>  1 12.8  -3.92  4.48  7.18
#>  2 -2.20  6.11 -2.16 -6.63
#>  3 17.5   2.15 18.7  -7.70
#>  4  3.35 12.9  18.0  19.4 
#>  5  6.65 -5.65 18.5  20.0 
#>  6 -4.28  7.74 15.5  10.0 
#>  7 20.9  -1.44  8.48 -2.83
#>  8 23.8  17.3  -5.62 14.8 
#>  9 -7.48  6.33 13.6  18.3 
#> 10 24.5  20.8  11.4  22.3 
#> # ℹ 15 more rows