Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                  X2       X3                  X4               X5   
#>    <dttm>              <time>   <dttm>              <chr>            <fct>
#>  1 2013-09-07 17:26:55 03:54:43 2002-11-20 01:50:53 HIiXWlBYRs3i9V   nume…
#>  2 2016-01-15 10:13:44 12:30:10 2005-09-16 12:28:43 NLnalLR          grum…
#>  3 2009-05-17 18:04:03 15:10:04 2018-11-20 13:49:53 BMgQ3SJOA5kMgmz… nume…
#>  4 2007-11-13 05:34:23 02:11:11 2005-11-16 21:50:47 Yep9vEK5         nume…
#>  5 2011-09-12 19:44:47 12:49:55 2003-02-11 02:39:36 BsR6ah7NBl9vc    hand…
#>  6 2018-10-07 01:29:30 07:21:43 2011-02-01 07:47:35 DwHGxp           kind…
#>  7 2007-05-04 20:19:22 15:35:29 2008-07-07 13:14:25 QVNtas6M5grAzX7… low_…
#>  8 2012-04-05 17:00:00 16:00:58 2002-08-23 04:45:39 Dhr26kv5Sg9cXXT… scar…
#>  9 2003-06-03 23:50:54 02:24:20 2016-09-19 02:31:24 cmwo11rspqhprS5p low_…
#> 10 2009-10-29 19:55:24 02:19:54 2020-08-28 06:20:32 LRaTY7C392       gray…

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2     X3     X4
#>      <dbl>   <dbl>  <dbl>  <dbl>
#>  1 -0.255   1.12    1.48  -0.383
#>  2 -0.309  -0.0747  0.219  0.584
#>  3 -1.71   -0.420  -0.970  0.963
#>  4 -0.753  -0.166  -1.40   2.56 
#>  5 -1.04   -0.0845  0.181  0.443
#>  6  1.06   -0.0827 -1.39  -0.805
#>  7  0.0678 -0.202  -0.958  0.870
#>  8  0.299  -0.743   0.928 -0.291
#>  9 -0.901   0.527  -0.447  0.380
#> 10 -0.577   0.475   1.21  -1.19 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2    X3    X4
#>    <dbl>  <dbl> <dbl> <dbl>
#>  1  5.48 19.8   11.7  21.9 
#>  2 21.4   5.88   2.73 13.6 
#>  3 13.1  16.5   -3.91 10.0 
#>  4 -7.98  2.00  -7.74 -4.23
#>  5 21.8   0.577 -5.82  5.57
#>  6  3.36 17.2   -6.66 13.2 
#>  7  3.66 14.0   18.1  -6.91
#>  8  1.64 17.2    2.13  6.67
#>  9 -4.57  9.16   8.92 16.2 
#> 10 23.8  15.3   -7.44  4.19
#> # … with 15 more rows