Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                      X2     X3      X4 X5                   
#>    <dttm>               <dbl>  <dbl>   <dbl> <chr>                
#>  1 2008-07-07 01:03:47  0.551 -1.55   0.0342 dVXJUFaL3FKAP5QvOVG  
#>  2 2011-01-15 01:53:56 -1.18   0.147 -1.32   HwADiD               
#>  3 2002-05-24 01:31:08  0.421 -0.544  1.07   2O9Z1Dzz0j75np1      
#>  4 2013-09-29 23:41:29 -1.29  -0.284  0.428  BUuCwmBndc3i         
#>  5 2018-04-02 18:12:56 -0.597 -0.390 -0.108  NJdLCyR94            
#>  6 2007-06-10 15:40:09  0.287 -1.13  -0.0521 dVaWlndD3VV6O        
#>  7 2012-01-19 04:22:30 -0.822  1.84  -0.300  iyak1G00LhRzYbY0ASQu 
#>  8 2017-04-30 15:40:14  0.313 -2.49   1.18   y9uoCR87Gyylc1QGgTw6h
#>  9 2006-02-22 21:21:44 -0.815 -2.24   1.15   sU0cAGx              
#> 10 2013-11-17 17:36:07  0.979  0.440  1.25   NAYxtcoO1WXStx0k0    

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2     X3      X4
#>      <dbl>   <dbl>  <dbl>   <dbl>
#>  1 -0.123  -0.502   0.452 -0.913 
#>  2  0.0496  1.22   -1.47   0.514 
#>  3  0.747  -0.877  -1.39  -0.776 
#>  4 -0.139  -0.413   0.689 -0.660 
#>  5 -0.844  -0.564   1.25   0.799 
#>  6 -0.0369 -0.365  -0.186 -0.701 
#>  7  0.486  -0.625  -1.91   1.62  
#>  8  2.19   -1.30    0.440  1.26  
#>  9 -1.75   -0.0717 -1.05   2.02  
#> 10 -0.0471  0.455   0.559 -0.0542
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1    X2    X3    X4
#>    <dbl> <dbl> <dbl> <dbl>
#>  1 -8.10 -8.66 18.4   5.03
#>  2  8.73 -4.79 21.9  -8.58
#>  3 14.5  22.7   3.80  2.22
#>  4 19.6  11.5  -7.08 -7.98
#>  5 -1.30 -1.66 -2.92  7.98
#>  6  7.78 20.0  -1.61 -8.13
#>  7 -8.83  4.93 -7.86 22.6 
#>  8  6.01 16.8  10.8  -3.89
#>  9  5.80 20.7  15.0  -5.82
#> 10 10.7  -2.75 11.5  -6.55
#> # … with 15 more rows