Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                  X2         X3         X4               X5        
#>    <dttm>              <date>     <date>     <fct>            <date>    
#>  1 2005-07-04 15:33:19 2007-04-14 2004-01-23 great_toad       2014-05-26
#>  2 2015-04-15 02:09:56 2015-05-15 2016-12-27 white_chimpanzee 2009-09-03
#>  3 2012-03-31 15:12:06 2019-12-29 2013-08-22 steep_cat        2018-10-23
#>  4 2015-10-30 12:03:21 2008-07-30 2013-02-07 huge_wildcat     2016-02-25
#>  5 2015-09-04 01:56:43 2017-02-11 2016-12-01 white_chimpanzee 2016-04-30
#>  6 2016-02-29 06:57:22 2011-12-06 2010-08-26 vast_okapi       2005-12-25
#>  7 2017-01-02 04:31:20 2014-02-10 2003-06-15 vast_okapi       2013-10-07
#>  8 2002-12-22 10:55:30 2015-08-05 2003-12-17 old_camel        2016-08-16
#>  9 2018-01-11 23:22:51 2018-05-26 2015-02-12 white_chimpanzee 2010-05-20
#> 10 2010-07-10 13:23:17 2009-05-08 2008-03-21 nervous_peccary  2001-12-09

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1     X2     X3      X4
#>     <dbl>  <dbl>  <dbl>   <dbl>
#>  1 -1.52  -0.482 -0.563  0.422 
#>  2  1.91   0.873  1.12  -1.94  
#>  3  0.879  0.267 -0.209 -0.0870
#>  4 -0.452 -1.82  -0.276  2.34  
#>  5  0.470  1.22  -1.50   3.29  
#>  6 -1.44  -0.458  2.56   0.189 
#>  7  0.502  1.24   0.392  1.82  
#>  8  1.45   0.678 -0.915 -1.09  
#>  9 -1.09   1.63  -0.620 -0.285 
#> 10  0.737 -0.547  0.831 -1.51  
#> # ℹ 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1     X2    X3    X4
#>     <dbl>  <dbl> <dbl> <dbl>
#>  1 24.5    0.604 -6.51 22.6 
#>  2 19.3   -1.74   8.28  5.20
#>  3 22.6   18.9   13.6  13.7 
#>  4 11.2   -2.94   5.82 10.9 
#>  5 16.6   13.5   24.9  12.2 
#>  6  9.77  17.2   -5.58  5.95
#>  7 21.8    3.31  11.2  13.3 
#>  8 15.6   -0.151  3.86  5.27
#>  9 -0.621 23.7    5.13  1.41
#> 10  8.27   2.28  23.0  18.3 
#> # ℹ 15 more rows