Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1         X2       X3                       X4      X5
#>    <date>     <time>   <fct>                 <int>   <dbl>
#>  1 2007-06-03 05:08:14 small_boar        315004467  1.35  
#>  2 2002-10-06 17:01:19 juicy_mole        664568115 -0.905 
#>  3 2005-05-14 14:55:04 thoughtless_toad 1421587447 -0.727 
#>  4 2009-04-17 06:31:33 thoughtless_toad 1123705042  0.868 
#>  5 2010-09-07 01:19:53 hissing_whale    2041312423 -0.852 
#>  6 2016-02-27 00:35:58 old_cougar       1369572570 -0.0582
#>  7 2020-09-09 19:31:25 grumpy_cougar     279943545 -0.726 
#>  8 2014-09-16 20:20:51 angry_ape        1362855064  0.982 
#>  9 2001-11-23 21:20:19 rich_crab        1858654146 -0.640 
#> 10 2013-02-17 10:23:17 grumpy_cougar     696244027 -0.853 

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1     X2     X3      X4
#>     <dbl>  <dbl>  <dbl>   <dbl>
#>  1 -0.696  0.485 -0.402 -2.26  
#>  2 -1.88   0.646 -2.20  -0.812 
#>  3 -0.711 -0.772 -1.19  -1.80  
#>  4 -0.792  1.19  -1.20   0.344 
#>  5 -0.437  0.936  1.01  -1.10  
#>  6 -0.344  0.807 -1.24  -0.436 
#>  7 -0.524 -2.01  -0.368 -0.965 
#>  8  0.650  0.300 -1.17  -0.990 
#>  9  1.27   0.629 -0.151  0.0463
#> 10 -0.361 -1.07   1.77   0.553 
#> # ℹ 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1    X2      X3    X4
#>    <dbl> <dbl>   <dbl> <dbl>
#>  1 -5.68 15.3   6.62   17.0 
#>  2  1.19  2.86  1.58   16.4 
#>  3  9.58  3.48 18.6    13.4 
#>  4  6.02 -5.10 -2.88   22.4 
#>  5  5.91  7.83 10.3    16.6 
#>  6  6.07 10.9  -2.14   -7.10
#>  7 -7.98  3.33 -4.89    3.81
#>  8  7.24  9.87  0.729  24.3 
#>  9 10.4   8.16  0.0451 20.0 
#> 10 23.6  12.2   5.63   -4.96
#> # ℹ 15 more rows