Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1                     X2         X3         X4               X5
#>    <chr>                  <date>     <date>     <time>        <int>
#>  1 vSd7vte                2011-01-05 2007-01-07 08:41:36 1548668178
#>  2 Ls0TuADDxnSNjMki8pB3x  2008-12-23 2008-04-20 18:04:41 1991018479
#>  3 FaHzaJ8QWegeB9wwd      2008-08-26 2011-08-15 16:32:36 2100713040
#>  4 CCVSzcaBrXFp9Q8U7m0vpX 2015-06-16 2013-01-06 05:35:56 1253078031
#>  5 PVV2C                  2018-09-08 2009-11-10 04:10:04 1570412969
#>  6 fwgZXtF8               2010-07-19 2014-07-09 10:26:29 1136791001
#>  7 9ZSRFBf                2011-02-22 2012-05-31 17:53:46 1605480567
#>  8 ayiVH                  2019-09-25 2017-06-25 00:57:09 1959818957
#>  9 Tly6Z                  2017-04-26 2001-09-02 21:40:35 1767408067
#> 10 gqzyB0Pmo86IN8mXio     2010-05-02 2009-05-06 07:06:48 1962142731

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1     X2       X3       X4
#>      <dbl>  <dbl>    <dbl>    <dbl>
#>  1 -1.20    1.43  -0.00249  1.39   
#>  2  0.842  -0.572  0.908    0.523  
#>  3 -1.43   -1.29  -0.673   -1.28   
#>  4 -0.883   0.243  0.452    1.17   
#>  5  0.236   0.150  0.305   -0.00409
#>  6  0.319   0.917 -2.25    -0.432  
#>  7  0.0656 -0.969  0.776   -0.261  
#>  8  1.22    1.54   0.267    0.142  
#>  9  2.28    0.382 -1.26     0.759  
#> 10 -0.147  -1.33  -0.458   -0.0879 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1      X2     X3    X4
#>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 -3.85  24.8    24.3   17.9 
#>  2 -8.80  -8.24   19.1    2.24
#>  3 24.9    0.0547 -0.508 -7.06
#>  4 21.4   -7.26   20.7   15.3 
#>  5 19.3   16.7    16.8   16.8 
#>  6 11.6   -4.98   -2.82   5.57
#>  7 -8.73   9.14   -0.514 22.6 
#>  8  0.215 -8.65   -2.64  -5.51
#>  9 16.8   -8.25    1.51   7.37
#> 10 23.6   20.1    12.2    7.74
#> # … with 15 more rows