Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>            X1     X2 X3                  X4             X5
#>         <int>  <dbl> <dttm>              <date>      <dbl>
#>  1  168963235  2.71  2020-04-19 10:04:09 2019-01-26  0.696
#>  2   77053786 -1.05  2008-01-07 02:46:35 2004-07-28  0.502
#>  3 1180023484  0.734 2011-07-13 00:58:14 2002-06-04  1.40 
#>  4  194871492  1.42  2014-12-24 19:53:42 2003-02-22  1.38 
#>  5  328909582  0.878 2006-03-12 04:49:22 2006-07-27  1.42 
#>  6  412771547 -1.52  2014-06-16 06:31:53 2012-02-26 -0.310
#>  7   91152255  1.64  2009-03-27 20:39:34 2008-05-29  0.490
#>  8 2009411851 -0.698 2019-09-17 18:37:25 2006-02-12  0.348
#>  9  853910930 -1.55  2016-10-07 02:09:32 2005-04-29  0.396
#> 10 2085534404 -0.913 2015-02-03 05:56:29 2015-02-21  0.387

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1      X2      X3     X4
#>      <dbl>   <dbl>   <dbl>  <dbl>
#>  1  0.748  -1.23    0.360  -1.83 
#>  2 -1.53    0.684  -0.986   0.351
#>  3 -1.03   -0.816   0.135   1.67 
#>  4 -1.41    0.312   2.17    0.361
#>  5  0.447  -0.285   0.379  -0.375
#>  6  0.390   1.62    0.0726 -0.136
#>  7  0.519   0.318  -2.53    0.207
#>  8  0.0980 -0.770   0.527  -1.08 
#>  9 -0.716   0.349   1.96    1.76 
#> 10  1.61   -0.0378 -0.870  -1.58 
#> # ℹ 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1     X2     X3     X4
#>     <dbl>  <dbl>  <dbl>  <dbl>
#>  1 -0.328  5.09  -4.27  14.6  
#>  2 19.9    9.11   2.23  -0.851
#>  3 23.6   22.9   12.0    6.59 
#>  4  1.32  -0.288 -0.350  9.84 
#>  5 18.8   22.6   16.9    0.766
#>  6 10.4   -9.87   2.03  -9.08 
#>  7 23.5    8.80  19.6   10.9  
#>  8  7.20  -8.92  19.6    5.67 
#>  9 16.3   17.4   13.2    2.32 
#> 10 -6.17   8.20   2.74  13.7  
#> # ℹ 15 more rows