Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>         X1 X2             X3                  X4                 X5
#>      <dbl> <fct>          <dttm>              <date>          <int>
#>  1  0.0804 filthy_gorilla 2004-10-06 05:31:23 2019-10-07   58986248
#>  2  0.274  filthy_gorilla 2016-11-19 16:17:43 2016-01-23  276608577
#>  3  0.702  filthy_gorilla 2001-06-25 14:22:40 2009-09-04  255483351
#>  4  1.33   filthy_gorilla 2014-05-26 15:50:06 2017-11-16  825334107
#>  5  1.18   filthy_gorilla 2014-06-10 08:24:13 2013-03-17 1247132296
#>  6  1.22   filthy_gorilla 2017-10-31 16:27:56 2002-06-18 1767980661
#>  7 -0.609  filthy_gorilla 2017-10-11 09:21:17 2013-06-18 1992365585
#>  8 -0.0339 filthy_gorilla 2003-11-17 20:46:53 2007-07-13  190112209
#>  9 -0.314  filthy_gorilla 2010-05-19 17:43:51 2008-08-04 1680216954
#> 10 -1.69   filthy_gorilla 2009-05-28 20:01:27 2001-03-20  638741953

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1     X2       X3      X4
#>      <dbl>  <dbl>    <dbl>   <dbl>
#>  1 -1.24   -0.124  0.355    1.17  
#>  2  0.114   0.468 -0.172   -0.708 
#>  3  0.692   0.322 -0.877    0.695 
#>  4 -0.682   0.284  2.70    -0.464 
#>  5 -0.153  -0.438  0.697   -0.175 
#>  6 -0.402   1.54   0.133    0.943 
#>  7 -0.0818 -1.20  -2.47    -0.840 
#>  8  1.39   -0.892 -0.00740  0.734 
#>  9  1.47   -0.108 -0.788    0.431 
#> 10  1.22   -0.634 -1.54    -0.0959
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1     X2    X3     X4
#>     <dbl>  <dbl> <dbl>  <dbl>
#>  1 12.0   -0.269  5.46  9.93 
#>  2 23.8   -7.39  -6.14 17.5  
#>  3  9.54  23.1   -8.92  5.23 
#>  4  3.81  -5.95  -7.23 18.1  
#>  5 23.9   14.0   10.6  -9.81 
#>  6  0.790 -9.40   9.65 15.0  
#>  7 -3.58  12.6   -3.77 -0.429
#>  8 -4.36  16.6   10.2  15.1  
#>  9  9.45  -9.15  -6.44  5.45 
#> 10  0.131 21.9   -1.85 19.6  
#> # … with 15 more rows