Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>         X1 X2         X3             X4                       X5        
#>      <dbl> <date>     <fct>          <chr>                    <date>    
#>  1 -0.777  2016-12-26 next_mare      f4XBQLts6fyvZe1Q8uka5vT  2013-04-02
#>  2 -0.255  2012-01-30 sweet_rat      IlHukJy                  2002-12-26
#>  3  1.59   2014-11-08 sweet_rat      FTH8yyEBHkpYn6qA9Oc6IiTb 2001-08-18
#>  4 -0.404  2015-08-02 swift_cat      4xkMxZQ                  2014-03-24
#>  5  0.0520 2005-03-10 last_ferret    gy4KGBwSnKDJtfc8         2018-07-23
#>  6  0.0250 2012-08-28 short_owl      ZKevsRdkD3H3TgzDTB       2014-08-28
#>  7 -0.375  2010-04-01 short_owl      Vmec84iAso3JHSC0         2020-04-04
#>  8  0.874  2008-02-23 deafening_seal 4h4e2F                   2019-10-27
#>  9  0.818  2010-03-15 hissing_okapi  wP0XA9rj2mRvaQKcxLt2F    2015-01-31
#> 10  0.0908 2004-11-03 short_owl      JacURNchJQFRBE           2018-05-26

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1      X2      X3     X4
#>     <dbl>   <dbl>   <dbl>  <dbl>
#>  1  0.863 -1.73    2.02    1.99 
#>  2  0.363  0.232  -0.724  -0.748
#>  3 -0.426  0.694  -0.0541  0.116
#>  4 -1.21   0.704   1.07    0.912
#>  5  2.59  -0.0668 -0.700   1.29 
#>  6 -0.313  1.12   -1.81    0.481
#>  7 -1.50   2.38    0.707   0.359
#>  8  0.348 -2.38    0.958  -0.186
#>  9  0.506  1.59    1.05    0.824
#> 10 -0.651  0.914   1.01    2.25 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1     X2     X3     X4
#>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1 21.8  -6.63  -4.65   9.65 
#>  2 17.1   5.63  22.8   -3.52 
#>  3 -5.07 17.9   20.8    9.95 
#>  4 18.6  21.9   22.1    9.47 
#>  5  1.65 20.8   -0.542  1.42 
#>  6  1.35  0.924 15.9   -9.65 
#>  7  3.96  9.10  18.2   20.1  
#>  8 -4.16 -6.90   7.83  10.7  
#>  9  2.63 -2.39  22.3   -5.76 
#> 10 -6.87 10.7    8.79   0.268
#> # … with 15 more rows