Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>            X1 X2         X3                  X4            X5
#>         <int> <date>     <dttm>              <time>     <dbl>
#>  1  351423299 2017-02-17 2019-10-04 02:36:31 05:05:52  0.120 
#>  2 1612200775 2004-12-09 2010-04-26 10:22:05 06:56:51  0.184 
#>  3  886712019 2006-04-13 2008-10-01 13:36:48 19:03:43  1.54  
#>  4  627946958 2016-12-09 2007-02-27 21:17:12 06:05:18  1.28  
#>  5  156196731 2003-09-28 2003-04-29 10:35:35 12:10:30 -0.655 
#>  6  962566647 2015-04-09 2004-11-10 08:44:39 22:01:50  1.19  
#>  7 1877385048 2005-04-10 2017-02-07 22:46:32 11:27:31 -2.57  
#>  8 1229416086 2011-09-14 2020-03-07 08:15:32 11:45:57 -0.456 
#>  9  930971813 2019-10-20 2010-04-13 09:26:13 20:26:49  0.521 
#> 10  626077312 2009-10-19 2011-08-10 00:48:39 02:24:13 -0.0628

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>        X1      X2     X3     X4
#>     <dbl>   <dbl>  <dbl>  <dbl>
#>  1  1.46  -1.22    0.547 -1.58 
#>  2 -0.505  0.0767 -1.24  -1.48 
#>  3  0.226 -0.389  -1.13  -0.768
#>  4 -0.688  0.644   0.414 -0.686
#>  5  0.123  0.0389 -1.16  -1.17 
#>  6 -0.865  0.0911  0.352 -0.655
#>  7 -0.675 -1.09    0.364 -0.772
#>  8  0.109  0.907  -0.366  2.68 
#>  9 -0.236  0.0895  0.145 -0.825
#> 10  1.61  -0.864   0.607 -1.41 
#> # ℹ 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>       X1    X2    X3    X4
#>    <dbl> <dbl> <dbl> <dbl>
#>  1 -3.87 15.9  -1.12 14.2 
#>  2 16.7  12.2  17.7   6.31
#>  3 15.7  -6.48 -9.53 13.6 
#>  4 -9.65  8.13 22.5  -4.61
#>  5 16.3  -8.35 24.1  11.7 
#>  6 18.2   8.66  4.65 -3.63
#>  7 -2.18 -5.67  7.58 21.9 
#>  8 22.1  -2.51 -5.59  4.58
#>  9 -9.48 14.2  22.0  21.1 
#> 10 12.8  13.4   9.69 18.6 
#> # ℹ 15 more rows