Generate a random tibble

This is useful for benchmarking, but also for bug reports when you cannot share the real dataset.

Usage

gen_tbl(
  rows,
  cols = NULL,
  col_types = NULL,
  locale = default_locale(),
  missing = 0
)

Arguments

rows

Number of rows to generate

cols

Number of columns to generate, if NULL this is derived from col_types.

col_types

One of NULL, a cols() specification, or a string.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

missing

The percentage (from 0 to 1) of missing data to use

Details

There is also a family of functions to generate individual vectors of each type.

See also

Examples

# random 10 x 5 table with random column types
rand_tbl <- gen_tbl(10, 5)
rand_tbl
#> # A tibble: 10 × 5
#>    X1            X2                  X3       X4       X5                 
#>    <chr>         <dttm>              <time>   <time>   <dttm>             
#>  1 cpLoXPBmmpZr… 2014-08-10 04:39:22 21:40:15 10:10:33 2001-11-27 12:41:39
#>  2 pcHTBKD4XuPY… 2003-02-25 13:11:18 15:27:56 10:08:50 2017-04-29 13:34:17
#>  3 JTYShjgYaffab 2019-10-14 03:20:20 19:32:48 01:34:52 2019-09-19 14:11:06
#>  4 ANoUa         2015-11-30 07:19:22 17:42:53 12:46:51 2019-05-23 03:10:25
#>  5 YeZVH9mGmmAG… 2001-03-06 05:43:33 05:42:23 01:06:01 2009-02-23 00:11:37
#>  6 2It0OSlQA0WX… 2016-05-14 05:29:01 19:00:49 07:34:46 2002-02-26 20:48:48
#>  7 c5MwE0caBXhL… 2009-08-14 03:18:28 16:36:50 08:39:14 2019-04-28 17:21:24
#>  8 zRuTAeJ3kU    2004-12-18 14:46:11 01:35:31 00:59:01 2003-05-12 04:07:38
#>  9 aojlIW        2020-07-25 20:15:17 08:48:12 13:50:19 2007-11-01 15:43:32
#> 10 zP97j83v7s0n… 2013-12-02 12:07:37 02:19:10 17:02:29 2017-11-04 15:47:25

# all double 25 x 4 table
dbl_tbl <- gen_tbl(25, 4, col_types = "dddd")
dbl_tbl
#> # A tibble: 25 × 4
#>         X1     X2     X3      X4
#>      <dbl>  <dbl>  <dbl>   <dbl>
#>  1  0.539   2.00  -1.29   0.982 
#>  2  1.56   -0.967 -0.794 -0.555 
#>  3  0.827  -1.41  -0.956 -0.575 
#>  4 -0.829   1.36   0.256  1.39  
#>  5  0.204  -0.860 -1.20   2.31  
#>  6 -0.155   0.793 -1.59   1.45  
#>  7 -1.39    0.187 -0.790 -0.546 
#>  8  0.0773 -0.418 -0.401 -0.809 
#>  9 -0.536   1.74   0.762 -0.0215
#> 10  0.664  -1.04   1.16   0.219 
#> # … with 15 more rows

# Use the dots in long form column types to change the random function and options
types <- rep(times = 4, list(col_double(f = stats::runif, min = -10, max = 25)))
types
#> [[1]]
#> <collector_double>
#> 
#> [[2]]
#> <collector_double>
#> 
#> [[3]]
#> <collector_double>
#> 
#> [[4]]
#> <collector_double>
#> 
dbl_tbl2 <- gen_tbl(25, 4, col_types = types)
dbl_tbl2
#> # A tibble: 25 × 4
#>        X1     X2    X3     X4
#>     <dbl>  <dbl> <dbl>  <dbl>
#>  1 -2.13  19.9    6.74 -0.143
#>  2  6.87   0.635 20.3  21.7  
#>  3  5.82   0.399 18.6   6.57 
#>  4 19.7   -8.07  -7.24 24.1  
#>  5 -5.47   8.38  -1.79 -8.18 
#>  6 -0.731  0.154  2.66 -6.68 
#>  7 24.5   24.8   -8.39 -2.67 
#>  8 -1.35   1.42  20.5  19.6  
#>  9 17.0   12.4    1.25 -3.69 
#> 10  8.58   8.56   3.35  8.94 
#> # … with 15 more rows