Generate individual vectors of the types supported by vroom

Generate individual vectors of the types supported by vroom

Usage,

gen_character(n, min = 5, max = 25, values = c(letters, LETTERS, 0:9), ...)

gen_double(n, f = stats::rnorm, ...)

gen_number(n, f = stats::rnorm, ...)

gen_integer(n, min = 1L, max = .Machine$integer.max, prob = NULL, ...)

gen_factor(
  n,
  levels = NULL,
  ordered = FALSE,
  num_levels = gen_integer(1L, 1L, 25L),
  ...
)

gen_time(n, min = 0, max = hms::hms(days = 1), fractional = FALSE, ...)

gen_date(n, min = as.Date("2001-01-01"), max = as.Date("2021-01-01"), ...)

gen_datetime(
  n,
  min = as.POSIXct("2001-01-01"),
  max = as.POSIXct("2021-01-01"),
  tz = "UTC",
  ...
)

gen_logical(n, ...)

gen_name(n)

Arguments

n

The size of the vector to generate

min

The minimum range for the vector

max

The maximum range for the vector

values

The explicit values to use.

...

Additional arguments passed to internal generation functions

f

The random function to use.

prob

a vector of probability weights for obtaining the elements of the vector being sampled.

levels

The explicit levels to use, if NULL random levels are generated using gen_name().

ordered

Should the factors be ordered factors?

num_levels

The number of factor levels to generate

fractional

Whether to generate times with fractional seconds

tz

The timezone to use for dates

Examples

# characters
gen_character(4)
#> [1] "AN5FKLYYcbE7nUa89wLb3AKrp" "Ciy4NfyFvUZGCFxIn8bJUhxq3"
#> [3] "HR6cHkv"                   "amFGN1TbM5b"              

# factors
gen_factor(4)
#> [1] scary_zebra        long_mongoose      gigantic_waterbuck gigantic_waterbuck
#> 15 Levels: obnoxious_seal screeching_bird dry_pony ... gigantic_waterbuck

# logical
gen_logical(4)
#> [1]  TRUE FALSE  TRUE  TRUE

# numbers
gen_double(4)
#> [1]  1.2393785  1.1432374 -1.7818473 -0.7591661
gen_integer(4)
#> [1] 1420160458  291838970 1969676101  612401386

# temporal data
gen_time(4)
#> 22:49:55
#> 06:11:47
#> 11:00:39
#> 12:02:45
gen_date(4)
#> [1] "2020-11-13" "2019-10-24" "2014-08-15" "2009-05-29"
gen_datetime(4)
#> [1] "2018-04-05 12:29:53 UTC" "2008-06-26 15:43:03 UTC"
#> [3] "2004-06-22 21:55:09 UTC" "2004-07-21 04:41:13 UTC"