Generate individual vectors of the types supported by vroom

Generate individual vectors of the types supported by vroom

Usage

gen_character(n, min = 5, max = 25, values = c(letters, LETTERS, 0:9), ...)

gen_double(n, f = stats::rnorm, ...)

gen_number(n, f = stats::rnorm, ...)

gen_integer(n, min = 1L, max = .Machine$integer.max, prob = NULL, ...)

gen_factor(
  n,
  levels = NULL,
  ordered = FALSE,
  num_levels = gen_integer(1L, 1L, 25L),
  ...
)

gen_time(n, min = 0, max = hms::hms(days = 1), fractional = FALSE, ...)

gen_date(n, min = as.Date("2001-01-01"), max = as.Date("2021-01-01"), ...)

gen_datetime(
  n,
  min = as.POSIXct("2001-01-01"),
  max = as.POSIXct("2021-01-01"),
  tz = "UTC",
  ...
)

gen_logical(n, ...)

gen_name(n)

Arguments

n

The size of the vector to generate

min

The minimum range for the vector

max

The maximum range for the vector

values

The explicit values to use.

...

Additional arguments passed to internal generation functions

f

The random function to use.

prob

a vector of probability weights for obtaining the elements of the vector being sampled.

levels

The explicit levels to use, if NULL random levels are generated using gen_name().

ordered

Should the factors be ordered factors?

num_levels

The number of factor levels to generate

fractional

Whether to generate times with fractional seconds

tz

The timezone to use for dates

Examples

# characters
gen_character(4)
#> [1] "mLOtLJ6nj4g1bMdsoHdtpIlM" "iUNK3VhE0IiDrD"          
#> [3] "IFyJDjRzItSsXwJ6f"        "AeQuqy7zOCfy949mYr1aY7k" 

# factors
gen_factor(4)
#> [1] dead_rat dead_rat dead_rat dead_rat
#> Levels: dead_rat

# logical
gen_logical(4)
#> [1] FALSE FALSE  TRUE  TRUE

# numbers
gen_double(4)
#> [1] -1.50150326  1.31418793 -0.09287389  0.92530891
gen_integer(4)
#> [1]  987006331  719777441  191854745 1304606468

# temporal data
gen_time(4)
#> 12:47:06
#> 00:05:39
#> 21:29:34
#> 12:39:02
gen_date(4)
#> [1] "2010-10-19" "2008-02-18" "2008-11-04" "2003-08-22"
gen_datetime(4)
#> [1] "2013-12-18 09:29:23 UTC" "2014-09-29 11:14:52 UTC"
#> [3] "2003-10-26 09:51:50 UTC" "2004-03-12 07:58:53 UTC"