Read a delimited file into a tibble

Read a delimited file into a tibble

vroom(
  file,
  delim = NULL,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  skip = 0,
  n_max = Inf,
  na = c("", "NA"),
  quote = "\"",
  comment = "",
  trim_ws = TRUE,
  escape_double = TRUE,
  escape_backslash = FALSE,
  locale = default_locale(),
  guess_max = 100,
  altrep = TRUE,
  altrep_opts = deprecated(),
  num_threads = vroom_threads(),
  progress = vroom_progress(),
  .name_repair = "unique"
)

Arguments

file
path to a local file.

delim
One or more characters used to delimit fields within a file. If NULL the delimiter is guessed from the set of c(",", "\t", " ", "|", ":", ";").

col_names
Either TRUE, FALSE or a character vector of column names.

If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc.

If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the output data frame.

Missing (NA) column names will generate a warning, and be filled in with dummy names X1, X2 etc. Duplicate column names will generate a warning and be made unique with a numeric prefix.

col_types
One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself.

If a column specification created by cols(), it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column: c = character, i = integer, n = number, d = double, l = logical, f = factor, D = date, T = date time, t = time, ? = guess, or _/- to skip the column.

col_select
One or more selection expressions, like in dplyr::select(). Use c() or list() to use more than one expression. See ?dplyr::select for details on available selection options.

id
Either a string or 'NULL'. If a string, the output will contain a variable with that name with the filename(s) as the value. If 'NULL', the default, no variable will be created.

skip
Number of lines to skip before reading data.

n_max
Maximum number of records to read.

na
Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

quote
Single character used to quote strings.

comment
A string used to identify comments. Any text after the comment characters will be silently ignored.

trim_ws
Should leading and trailing whitespace be trimmed from each field before parsing it?

escape_double
Does the file escape quotes by doubling them? i.e. If this option is TRUE, the value '""' represents a single quote, '"'.

escape_backslash
Does the file use backslashes to escape special characters? This is more general than escape_double as backslashes can be used to escape the delimiter character, the quote character, or to add special characters like \n.

locale
The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

guess_max
Maximum number of records to use for guessing column types.

altrep
Control which column types use Altrep representations, either a character vector of types, TRUE or FALSE. See vroom_altrep() for for full details.

altrep_opts

num_threads
Number of threads to use when reading and materializing vectors. If your data contains embedded newlines (newlines within fields) you must use num_threads = 1 to read the data properly.

progress
Display a progress bar? By default it will only display in an interactive session and not while knitting a document. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more. The automatic progress bar can be disabled by setting option readr.show_progress to FALSE.

.name_repair
Handling of column names. By default, vroom ensures column names are not empty and unique. See .name_repair as documented in tibble::tibble() for additional options including supplying user defined name repair functions.

Examples

# Show path to example file
input_file <- vroom_example("mtcars.csv")

# Read from a path

# Input sources -------------------------------------------------------------
# Read from a path
vroom(input_file)

#> Rows: 32
#> Columns: 12
#> Delimiter: ","
#> chr [ 1]: model
#> dbl [11]: mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 32 x 12
#>    model         mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <chr>       <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Mazda RX4    21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2 Mazda RX4 …  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3 Datsun 710   22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4 Hornet 4 D…  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5 Hornet Spo…  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6 Valiant      18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7 Duster 360   14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8 Merc 240D    24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9 Merc 230     22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10 Merc 280     19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows

# You can also use literal paths directly
# vroom("mtcars.csv")

if (FALSE) {
# Including remote paths
vroom("https://github.com/r-lib/vroom/raw/master/inst/extdata/mtcars.csv")
}

# Or directly from a string (must contain a trailing newline)
vroom("x,y\n1,2\n3,4\n")

#> Rows: 2
#> Columns: 2
#> Delimiter: ","
#> dbl [2]: x, y
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 2 x 2
#>       x     y
#>   <dbl> <dbl>
#> 1     1     2
#> 2     3     4

# Column selection ----------------------------------------------------------
# Pass column names or indexes directly to select them
vroom(input_file, col_select = c(model, cyl, gear))

#> Rows: 32
#> Columns: 3
#> Delimiter: ","
#> chr [1]: model
#> dbl [2]: cyl, gear
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 32 x 3
#>    model               cyl  gear
#>    <chr>             <dbl> <dbl>
#>  1 Mazda RX4             6     4
#>  2 Mazda RX4 Wag         6     4
#>  3 Datsun 710            4     4
#>  4 Hornet 4 Drive        6     3
#>  5 Hornet Sportabout     8     3
#>  6 Valiant               6     3
#>  7 Duster 360            8     3
#>  8 Merc 240D             4     4
#>  9 Merc 230              4     4
#> 10 Merc 280              6     4
#> # … with 22 more rows

vroom(input_file, col_select = c(1, 3, 11))

#> Rows: 32
#> Columns: 3
#> Delimiter: ","
#> chr [1]: model
#> dbl [2]: cyl, gear
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 32 x 3
#>    model               cyl  gear
#>    <chr>             <dbl> <dbl>
#>  1 Mazda RX4             6     4
#>  2 Mazda RX4 Wag         6     4
#>  3 Datsun 710            4     4
#>  4 Hornet 4 Drive        6     3
#>  5 Hornet Sportabout     8     3
#>  6 Valiant               6     3
#>  7 Duster 360            8     3
#>  8 Merc 240D             4     4
#>  9 Merc 230              4     4
#> 10 Merc 280              6     4
#> # … with 22 more rows

# Or use the selection helpers
vroom(input_file, col_select = starts_with("d"))

#> Rows: 32
#> Columns: 2
#> Delimiter: ","
#> dbl [2]: disp, drat
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 32 x 2
#>     disp  drat
#>    <dbl> <dbl>
#>  1  160   3.9 
#>  2  160   3.9 
#>  3  108   3.85
#>  4  258   3.08
#>  5  360   3.15
#>  6  225   2.76
#>  7  360   3.21
#>  8  147.  3.69
#>  9  141.  3.92
#> 10  168.  3.92
#> # … with 22 more rows

# You can also rename specific columns
vroom(input_file, col_select = list(car = model, everything()))

#> Rows: 32
#> Columns: 12
#> Delimiter: ","
#> chr [ 1]: model
#> dbl [11]: mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 32 x 12
#>    car           mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <chr>       <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Mazda RX4    21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2 Mazda RX4 …  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3 Datsun 710   22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4 Hornet 4 D…  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5 Hornet Spo…  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6 Valiant      18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7 Duster 360   14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8 Merc 240D    24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9 Merc 230     22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10 Merc 280     19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows

# Column types --------------------------------------------------------------
# By default, vroom guesses the columns types, looking at 1000 rows
# throughout the dataset.
# You can specify them explcitly with a compact specification:
vroom("x,y\n1,2\n3,4\n", col_types = "dc")

#> # A tibble: 2 x 2
#>       x y    
#>   <dbl> <chr>
#> 1     1 2    
#> 2     3 4    

# Or with a list of column types:
vroom("x,y\n1,2\n3,4\n", col_types = list(col_double(), col_character()))

#> # A tibble: 2 x 2
#>       x y    
#>   <dbl> <chr>
#> 1     1 2    
#> 2     3 4    

# File types ----------------------------------------------------------------
# csv
vroom("a,b\n1.0,2.0\n", delim = ",")

#> Rows: 1
#> Columns: 2
#> Delimiter: ","
#> dbl [2]: a, b
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 1 x 2
#>       a     b
#>   <dbl> <dbl>
#> 1     1     2

# tsv
vroom("a\tb\n1.0\t2.0\n")

#> Rows: 1
#> Columns: 2
#> Delimiter: "\t"
#> dbl [2]: a, b
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 1 x 2
#>       a     b
#>   <dbl> <dbl>
#> 1     1     2

# Other delimiters
vroom("a|b\n1.0|2.0\n", delim = "|")

#> Rows: 1
#> Columns: 2
#> Delimiter: "|"
#> dbl [2]: a, b
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message

#> # A tibble: 1 x 2
#>       a     b
#>   <dbl> <dbl>
#> 1     1     2